+++ plugins/postcss-color-mod-function/index.js
@@ -1,15 +1,19 @@
-// tooling
+import { importCustomPropertiesFromSources, importCustomPropertiesFromCSSAST } from './lib/import';
 import parser from 'postcss-values-parser';
 import postcss from 'postcss';
 import transformAST from './lib/transform';
 
-// plugin
 export default postcss.plugin('postcss-color-mod-function', opts => {
 	const unresolvedOpt = String(Object(opts).unresolved || 'throw').toLowerCase();
 	const stringifierOpt = Object(opts).stringifier || (color => color.toLegacy());
+	const importFrom = [].concat(Object(opts).importFrom || []);
 	const transformVarsOpt = 'transformVars' in Object(opts) ? opts.transformVars : true;
 
-	return (root, result) => {
+	const customPropertiesPromise = importCustomPropertiesFromSources(importFrom);
+
+	return async (root, result) => {
+		const customProperties = Object.assign(await customPropertiesPromise, await importCustomPropertiesFromCSSAST(root));
+
 		root.walkDecls(decl => {
 			const originalValue = decl.value;
 
@@ -21,7 +25,8 @@ export default postcss.plugin('postcss-color-mod-function', opts => {
 					stringifier: stringifierOpt,
 					transformVars: transformVarsOpt,
 					decl,
-					result
+					result,
+					customProperties
 				});
 
 				const modifiedValue = ast.toString();