+++ plugins/postcss-color-mod-function/test/import-root.js
@@ -1,0 +1,7 @@
+module.exports = {
+	customProperties: {
+		'--color-blue': 'blue',
+		'--color-red': 'red',
+		'--color': 'var(--color-blue)'
+	}
+};